pyproject.toml
[project]
name = "embeddr-core"
version = "0.1.5"
description = "Embeddr Core Library"
readme = "README.md"
authors = [{ name = "Nynxz", email = "contact@nynxz.com" }]
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.14.10",
    "sqlmodel>=0.0.29",
    "pillow>=12.0.0",
    "numpy>=2.4.0",
    "transformers>=4.57.3",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
torch = ["torch>=2.9"]
dev = ["pytest", "ruff"]

[project.urls]
Homepage = "https://github.com/embeddr-net/embeddr-core"
Repository = "https://github.com/embeddr-net/embeddr-core"
Issues = "https://github.com/embeddr-net/embeddr-core/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.mypy]
check-untyped-defs = true

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
